value-pairs: Rename replace() to replace-prefix()
This renames the replace() value-pairs rekey option to
replace-prefix(), because that is what it does. This is done for both
the value-pairs() grammar, and for the command-line-like parser used
for template functions too.

For compatibility's sake, replace() remains as an alias for now, but
it is marked as obsolete, and will emit a warning.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>