libtest: Added AssertPipe for assert messages arriving on a LogPipe.
AssertPipe provides a LogPipe, on which we can assert on the nvpairs
of an arriving LogMessage.

Unfortunately, so far only one nvpair/LogMessage can be asserted.

Example:
  assert_pipe_new("prop1", "value1", "prop2", "value2");

In this example, the first message is checked if it's "prop1" nvpair
has value "value1", the second message is checked if it's "prop2" nvpair
has value "value2".

If more LogMessages arrives on the AssertPipe, than nvpairs specified in the
constructor, and assertion is raised.