affile: fixed a pipe permission problem
When a pipe is created by syslog-ng, after creation, it is opened with 0666
permissions. Permissions are apply at a later point, if the user set
for the pipe. Otherwise, 0666 will remain. But even the permissions are set in
the config, still there is a time-gap while the pipe can be accessed with 0666.

Fix: open the pipe with a more strict permission[0600], then, if exists,
we apply the permissions found in the config file.

Signed-off-by: Budai Laszlo <stentor.bgyk@gmail.com>